AMBARI-22441 Integrate Alert Groups on Alert page with websocket events. (atkach)

+++ ambari-web/app/mappers/alert_groups_mapper.js
@@ -39,7 +39,7 @@ App.alertGroupsMapper = App.QuickDataMapper.create({
     }
   },
 
-  map: function (json) {
+  map: function (json, ignoreDelete) {
     if(Em.isNone(App.cache['previousAlertGroupsFullMap'])) {
       App.cache['previousAlertGroupsFullMap'] = {};
     }
@@ -101,9 +101,11 @@ App.alertGroupsMapper = App.QuickDataMapper.create({
 
       }, this);
 
-      groupsToDelete.forEach(function(groupId) {
-        self.deleteRecord(App.AlertGroup.find(groupId));
-      });
+      if (!ignoreDelete) {
+        groupsToDelete.forEach(function(groupId) {
+          self.deleteRecord(App.AlertGroup.find(groupId));
+        });
+      }
 
       App.cache['previousAlertGroupsMap'] = alertDefinitionsGroupsMap;
       App.cache['previousAlertGroupsFullMap'] = groupsMap;
